third_party/WebKit/LayoutTests/fast/dom/HTMLAnchorElement/set-href-attribute-host-expected.txt

Test setting the host attribute of the URL in HTMLAnchorElement.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Basic test
PASS a.href is 'https://www.otherdomain.com:0/path/'
Set host without port
PASS a.href is 'https://www.otherdomain.com:8080/path/'
Set host with '?' in it
FAIL a.href should be https://www.other/?domain.com:8080/path/?key=value. Was https://www.other%3Fdomain.com:8080/path/?key=value.
Set default port for another protocol
PASS a.href is 'https://www.otherdomain.com:80/path/'
Set default port
PASS a.href is 'https://www.otherdomain.com/path/'
Set host with invalid port
PASS a.href is 'https://www.otherdomain.com:0/path/'
Set host with letters in port number
PASS a.href is 'https://www.otherdomain.com:44/path/'
Leading space in port number
PASS a.href is 'https://www.otherdomain.com:0/path/'
Colon without port number
PASS a.href is 'https://www.otherdomain.com:0/path/'
Set host to null
PASS a.href is 'https://null:8080/path/'
Set host to empty string
PASS a.href is 'https://www.mydomain.com:8080/path/'
Set host to URL with file: protocol
PASS a.href is 'file://mydomain.com/path/'
Set host containing slashes in it
PASS a.href is 'https://www.otherdom%2Fain.com:8080/path/'
Set host to a malformed URL
FAIL a.href should be https:/
ww.my@domain.com:8080/path/. Was https://ww.my@www.other%21domain.com:15/path/.
Set host that starts with ':'
PASS a.href is 'https://domain.com:8080/path/'
Set host to URL containing username and ..
FAIL a.href should be https://rwwmy@www.other!domain.com:25/pa..th/. Was https://rwwmy@www.other%21domain.com:25/pa..th/.
Set host to a URL with tel: protocol
PASS a.href is 'tel:+1-816-555-1212'
PASS successfullyParsed is true

TEST COMPLETE
